moved commands to their own files, added error if !accept is called on empty queue
